#209 Rename utility module
