+++ neclib/utils/__init__.py
@@ -1,7 +1,7 @@
 """Collection of general purpose functionalities."""
 
 from .data_utils import *  # noqa: F401, F403
-from .logic_utils import *  # noqa: F401, F403
+from .device_utils import *  # noqa: F401, F403
 from .math_utils import *  # noqa: F401, F403
 from .quantity_utils import *  # noqa: F401, F403
 from .simulator_utils import *  # noqa: F401, F403
